Update API of py::oset
