+++ src/core/frame/__init__.cc
@@ -25,9 +25,6 @@
 #include <vector>
 #include "column/npmasked.h"
 #include "python/_all.h"
-#include "python/list.h"
-#include "python/oset.h"
-#include "python/string.h"
 #include "utils/alloc.h"
 #include "stype.h"
 #include "ztest.h"
